+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add `lyp benchmark` command (#17).
 - Add shortcuts for common commands.
 - Add `lyp watch` command (#38).